# Conflicts:
#	src/main/java/com/firstgroup/gamemanagerapi/controller/UserController.java
#	src/main/java/com/firstgroup/gamemanagerapi/model/mapper/DeveloperMapper.java
#	src/main/java/com/firstgroup/gamemanagerapi/model/mapper/GameMapper.java
#	src/main/java/com/firstgroup/gamemanagerapi/model/mapper/GenreMapper.java
#	src/main/java/com/firstgroup/gamemanagerapi/repository/GenreRepository.java
#	src/main/java/com/firstgroup/gamemanagerapi/service/GenreService.java
#	src/main/java/com/firstgroup/gamemanagerapi/service/UserService.java
#	src/main/java/com/firstgroup/gamemanagerapi/util/ResponseUtils.java

+++ src/main/java/com/firstgroup/gamemanagerapi/service/UserService.java
@@ -1,5 +1,6 @@
 package com.firstgroup.gamemanagerapi.service;
 
+import com.firstgroup.gamemanagerapi.exception.ResourceNotFoundException;
 import com.firstgroup.gamemanagerapi.model.dto.UserDTO;
 import com.firstgroup.gamemanagerapi.model.entity.User;
 import com.firstgroup.gamemanagerapi.model.mapper.UserMapper;
@@ -18,6 +19,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
 @Slf4j
 @Service
@@ -31,15 +34,19 @@ public class UserService {
     private final UserRepository userRepository;
     private final UserMapper userMapper;
 
+
     @Transactional
     public UserDTO createUser(@Valid UserRO ro) {
-        User user = userMapper.toEntity(ro);
-        user.setPassword(ro.password());
-
         try {
-            return userMapper.toDto(userRepository.save(user));
-        } catch (DataIntegrityViolationException ex) {
-            throw new ResponseStatusException(HttpStatus.CONFLICT, "User with this display name or email already exists.");
+            User user = userMapper.toEntity(ro);
+            user.setPassword(ro.password());
+
+            User savedUser = userRepository.save(user);
+            log.info(MessageUtils.saveSuccessMessage(USER));
+
+            return userMapper.toDto(savedUser);
+        } catch (DataIntegrityViolationException e) {
+            throw new RuntimeException(MessageUtils.saveErrorMessage(USER));
         }
     }
 
@@ -55,19 +62,76 @@ public class UserService {
         }
     }
 
-    public UserDTO getUserById(long id) {
-        return userMapper.toDto(userRepository.findById(id)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User with this ID does not exist.")));
+    public Optional<User> getById(Long id) {
+        if (Objects.isNull(id)) {
+            return Optional.empty();
+        }
+
+        return userRepository.findById(id);
     }
 
-    public UserDTO getUserByDisplayName(String displayName) {
-        return userMapper.toDto(userRepository.findByDisplayName(displayName)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User with this display name does not exist.")));
+    public Optional<User> getByEmail(String email) {
+        if (Objects.isNull(email)) {
+            return Optional.empty();
+        }
+
+        return userRepository.findByEmail(email);
     }
 
-    public UserDTO getUserByEmail(String email) {
-        return userMapper.toDto(userRepository.findByEmail(email)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User with this email does not exist.")));
+    public Optional<User> getByDisplayName(String displayName) {
+        if (Objects.isNull(displayName)) {
+            return Optional.empty();
+        }
+
+        return userRepository.findByDisplayName(displayName);
+    }
+
+    public User getUserById(Long id) {
+        try {
+            Optional<User> user = getById(id);
+
+            if (user.isEmpty()) {
+                throw new Exception("User not found.");
+            }
+            log.info(MessageUtils.retrieveSuccessMessage(USER));
+            return user.get();
+        } catch (Exception e) {
+            String errorMessage = MessageUtils.retrieveErrorMessage(USER);
+            log.error(errorMessage);
+            throw new ServiceException(errorMessage, e);
+        }
+    }
+
+    public User getUserByEmail(String email) {
+        try {
+            Optional<User> user = getByEmail(email);
+
+            if (user.isEmpty()) {
+                throw new Exception("User not found.");
+            }
+            log.info(MessageUtils.retrieveSuccessMessage(USER));
+            return user.get();
+        } catch (Exception e) {
+            String errorMessage = MessageUtils.retrieveErrorMessage(USER);
+            log.error(errorMessage);
+            throw new ServiceException(errorMessage, e);
+        }
+    }
+
+    public User getUserByDisplayName(String displayName) {
+        try {
+            Optional<User> user = getByDisplayName(displayName);
+
+            if (user.isEmpty()) {
+                throw new Exception("User not found.");
+            }
+            log.info(MessageUtils.retrieveSuccessMessage(USER));
+            return user.get();
+        } catch (Exception e) {
+            String errorMessage = MessageUtils.retrieveErrorMessage(USER);
+            log.error(errorMessage);
+            throw new ServiceException(errorMessage, e);
+        }
     }
 
     @Transactional
@@ -99,8 +163,18 @@ public class UserService {
 
     @Transactional
     public void deleteUser(Long id) {
-        User user = userRepository.findById(id)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User with this ID does not exist."));
-        userRepository.delete(user);
+        try {
+            User user = getUserById(id);
+
+            if (Objects.isNull(user)) {
+                throw new ResourceNotFoundException("User not found");
+            }
+
+            userRepository.delete(user);
+        } catch (Exception e) {
+            String errorMessage = MessageUtils.deleteErrorMessage(USER);
+            log.error(errorMessage);
+            throw new ServiceException(errorMessage, e);
+        }
     }
 }
